final update

+++ src/main/scala/wcc/CommunityData.scala
@@ -8,7 +8,7 @@ package wcc
  * @param a number of internal edges.
  * @param b number of external edges.
  */
-class CommunityData(val r: Int, val a: Int, val b: Int) extends Serializable {
+class CommunityData(val r: Int, val a: Double, val b: Int) extends Serializable {
   // the edge density δ = 2 * number of edges / squared number of vertices.
   val d = 2 * a / math.pow(r, 2)
